Move Package Source Serializer to its own file
